KAS-wgpu: move texture atlas code to new module (from images)
